Major changes: custom user model
I've switched everything over to use a custom user model instead of the default django one. This was quite complex and required lots of changes to views, models and forms, but is worth it. It allows us to use email addresses for logins, make sure email addresses are unique and move all the email convenience functions over to the user model.  It also gives us more flexibility in the future.