Added sent to xero feature
A new contact will attempt to be created in Xero when any of the following happens:
1) The "Make Member" button is clicked in the member portal.
or 2) The "Resend To Xero" button is pressed.

In both instances, the system checks if the same contact name or email has been used before. It throws an error if it has. (you'll need to manually set the xero id/account in the db but this should never happen). It will also check if the account ID has been used before and attempt to generate one if necessary.